+++ src/index.ts
@@ -1,1 +1,2 @@
 export { Producer } from './producer';
+export * from './types';
